splitting utils.ts in two files
