+++ packages/core/src/utils/log.ts
@@ -1,0 +1,6 @@
+const generateStackTrace = (): string => {
+  const err = new Error();
+  return err.stack || "";
+};
+
+export {generateStackTrace};
